feat: ìopensearch 연결 완료ã
